backend/requirements.txt

Flask==2.2.3
Flask-Cors==3.0.10
rembg==2.0.50
Werkzeug==2.2.3
gunicorn==20.1.0
Pillow==9.5.0
numpy==1.24.3
